Setup nightly builds from copier_template
Based on copier_template@nightly2.

I am not happy about re-listing test and docs dependencies. Try problem
is that we cannot use the requirement filts: If we move, e.g., the docs
dependencies (minus the base.in) into its own file, pip-compile-multi
cannot solve, for example because base.in resolves a different version
of cloudpickle than the docs-only requirements.