Try to speed up tests by caching loaded files
There is probably a way to tweak the caching in the I(Q) beam-center
finder to fix the root cause, but modifying that code was already
required way too often. I'd like to avoid that for now. Ideally I would
like to use the new ess.reduce.streaming.StreamProcessor, but the
control flow is just too convoluted to make this easy.