Try sharing common parts of workflow compute to make tests run faster
This is demonstrating an approach that could save some test runtime. In
this example a parametrized test is running a nearly identical workflow
8 times, each time loading data from scratch, etc.

The change here moves some of the common parts into a module-scope
fixture. On my machine this reduces the test runtime from 50 seconds to
22 seconds.

I do not know if this is a great solution. If we want to go with this, a
helper utility could be made: Given a sciline.Pipeline and a set of
keys, compute all intermediate results that to not depend on those keys
and set the results as "static" data in the workflow. This would avoid
error-prone manual authoring of fixtures like the one I added here.